# week-4/aiWorker/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FlyRank AI Worker Auth API",
    "version": "1.0.0",
    "description": "Express API using Supabase Auth with JWT bearer authentication"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "tags": [
    { "name": "Auth" },
    { "name": "Public" },
    { "name": "Protected" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Credentials": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "format": "password", "minLength": 1 }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } }
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Create a Supabase user",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Credentials" } } }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Invalid input or Supabase signup error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Sign in and receive JWT bearer tokens",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Credentials" } } }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "400": { "description": "Missing credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Validate the current bearer token and log out",
        "description": "The client must discard its access and refresh tokens after this response.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logout successful" },
          "401": { "description": "Missing or invalid bearer token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/public/info": {
      "get": {
        "tags": ["Public"],
        "summary": "Get public information",
        "responses": { "200": { "description": "Public information" } }
      }
    },
    "/protected/profile": {
      "get": {
        "tags": ["Protected"],
        "summary": "Get the authenticated user's profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Authenticated user information" },
          "401": { "description": "Missing or invalid bearer token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "tags": ["Protected"],
        "summary": "Get the protected dashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Protected dashboard" },
          "401": { "description": "Missing or invalid bearer token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    }
  }
}